requirements.txt
langchain_openai
langchain_core
langchain_community 
streamlit 
python-dotenv
langchain 
langserve
fastapi 
uvicorn 
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
langchainhub
langchain_ollama
groq 
langchain_groq
sentence_transformers